Revert "Removing identity section and moving all related stuff to user-data section in auto-install file of Subiquity. Needed to set user group correctly"
The installer for Ubuntu 26.04 now has the ability to create users before first
boot. This will fix the inability to install unattended when Guest Additions
are also installed (see ticketref:22278).

But for that to work, we need the user parameters to be specified in the
identity section, not in the user-data section (which is processed by
cloud-init on first boot).

ticketref:22278

This reverts commit 911ce9ed60645a532aa8b6bb1557dbeac1dd2138.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>